feat(tui): add shared terminal foundation
