+++ crates/tui/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "foundry-tui"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+crossterm.workspace = true
+ratatui = { workspace = true, features = ["crossterm"] }
